+++ astroML/datasets/__init__.py
@@ -23,3 +23,4 @@ from .hogg2010test import fetch_hogg2010test
 from .rrlyrae_templates import fetch_rrlyrae_templates
 from .sdss_filters import fetch_sdss_filter, fetch_vega_spectrum
 from .kelly2007test import simulation_kelly
+from .sdss_galaxy_images import fetch_sdss_galaxy_images